Switch IMG_File::readImages() and the deallocating overload of IMG_File::writeImages() to use UT_Array<UT_UniquePtr<PXL_Raster>>.
Memory leaks related to these APIs have already been committed except for
modifications to QT_SnapshotGalleryModel::generateThumbnailAndCache_() which
are fixed as part of this commit.

As part of this change TIL_Raster's constructor which takes ownership of its an
array of images parameter has also been changed to take a
UT_Array<UT_UniquePtr<PXL_Raster>>.

The other notable change here is that GLTF_ImageExport was changed to use
UT_UniquePtr as well where it was previously using UT_SharedPtr. This lowers
the memory usage slightly and also fixes some memory fragmentation since it was
not using UTmakeShared() to allocate.